requirements.txt
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
functions-framework==3.5.0
pydantic==2.6.3
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
asgiref==3.7.2
flask==2.3.3
werkzeug==2.3.7
click==8.1.7
starlette==0.36.3 